Duong test model

+++ Client_WS_Supermarket/src/main/java/com/example/client_ws_supermarket/controller/ProductController.java
@@ -5,13 +5,8 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.client.RestTemplate;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
 
 @Controller
 @RequestMapping("product")
@@ -19,7 +14,7 @@ public class ProductController {
     protected RestTemplate rest = new RestTemplate();
 
     @GetMapping("{id}")
-    public String cartView(Model model, @PathVariable Long id) {
+    public String cartView(Model model, @PathVariable Integer id) {
         Product p = new Product();
 
         String url = "http://localhost:8081/product?id=" + id;
